refactor: moved method only used for testing inside testing object
