+++ engine/src/main/kotlin/io/github/agentseek/util/FastEntities.kt
@@ -15,43 +15,6 @@ import io.github.agentseek.world.World
 import kotlin.time.Duration
 
 object FastEntities {
-
-    /**
-     * Creates a component from a scene, returning the created GameObject and Component
-     */
-    fun Scene.createComponent(
-        initFun: Component.() -> Unit = {},
-        updateFun: Component.(Duration) -> Unit = {},
-        onRemovedFun: Component.() -> Unit = {},
-        addToGameObject: Boolean = true,
-        addToWorld: Boolean = true
-    ): Pair<GameObject, Component> {
-        val go = world.gameObjectBuilder().build().also {
-            if (addToWorld) {
-                world.addGameObject(it)
-            }
-        }
-        val component = object : AbstractComponent(go) {
-            override fun init() {
-                initFun()
-            }
-
-            override fun onUpdate(deltaTime: Duration) {
-                updateFun(deltaTime)
-            }
-
-            override fun onRemoved() {
-                onRemovedFun()
-            }
-        }.also {
-            if (addToGameObject) {
-                go.addComponent(it)
-            }
-        }
-
-        return Pair(go, component)
-    }
-
     /**
      * Creates an empty `GameObject` in the scene.
      *
